Mem class

Mem class containing results from multiEnrichMap().

Usage

# S4 method for class 'Mem,ANY,ANY,ANY'
x[i, j, ..., drop = TRUE]

list_to_Mem(mem)

Mem_to_list(x, ...)

# S4 method for class 'Mem'
enrichments(x)

# S4 method for class 'Mem'
names(x)

# S4 method for class 'Mem'
geneIM(x, ...)

# S4 method for class 'Mem'
enrichIM(x, ...)

# S4 method for class 'Mem'
memIM(x, ...)

# S4 method for class 'Mem'
genes(x, ...)

# S4 method for class 'Mem'
thresholds(x, ...)

# S4 method for class 'Mem'
sets(x, ...)

Arguments

x

Mem object

i

character or integer with gene set / pathway names

j

character or integer with enrichment names

...

additional arguments are ignored

mem

list output from multiEnrichMap()

Value

list_to_Mem() returns a Mem S4 object.

Mem_to_list() returns a list suitable for ther mem functions.

Details

Slots

enrichList: list of enrichResults named by enrichment test.

enrichLabels: character vector with custom enrichment labels, named by enrichment test.

colorV: character vector of R colors, named by enrichment test.

geneHitList: list of vectors named by enrichment test. Vectors are either:
character vector of genes tested as hits for enrichment, or
integer vector named by genes with values indicating the directionality, using 1 or -1.

geneHitIM: numeric matrix with enrichment tests as colnames, and genes as rownames.
Values are 1 or 0 to indicate gene hits tested,
optionally 1 or -1 to indicate directionality.

memIM: numeric matrix with gene sets / pathways as colnames, and genes as rownames.
Values are 1 or 0 indicating genes present in each pathway.

geneIM: numeric matrix with enrichment test as colnames, and genes as rownames.
This matrix differs from 'geneHitIM' in that it only includes genes also in 'memIM', and in that order.

enrichIM: numeric matrix with enrichment test as colnames, and gene sets / pathways as rownames.
Matrix values are enrichment P-values, using the appropriate adjusted P-value or FDR to represent significance.

multiEnrichDF: data.frame with single combined enrichment table.
Data include union of genes involved in each pathway, and lowest P-value observed across all enrichment tests.

multiEnrichResult: enrichResult equivalent of 'multiEnrichDF'.
It is intended to facilitate use in enrichplot functions.

thresholds: list of thresholds used when calling multiEnrichMap().

colnames: list to associate actual colnames in 'enrichList' with the conceptual names used in multienrichjam.

geneIMcolors: character matrix with enrichment test colnames, and gene rownames.
Values are colors assigned using 'colorV'.

geneIMdirection: numeric matrix equivalent to 'geneIM' with values 1 and -1 indicating directionality.
When no directionality is provided, all values will be 1 or 0.

enrichIMcolors: character matrix with enrichment test colnames, and gene set / pathway rownames.
Values are color gradients using the appropriate P-value threshold used with multiEnrichMap().

enrichIMdirection: numeric matrix equivalent to 'enrichIM' with values indicating directional score.
Directional scores may include a IPA z-score or other directional score where available.
When not available, all values are 1.

enrichIMgeneCount: integer matrix equivalent to 'enrichIM'.
Values are the number of genes involved in the enrichment of each pathway.

Functions

x[i: Subset a Mem object

list_to_Mem(): Convert list to Mem S4 object

Mem_to_list(): Coerce Mem object to legacy 'mem' list format

enrichments(Mem): Names for the enrichment tests in a Mem object

names(Mem): Names for the enrichment tests in a Mem object

geneIM(Mem): The matrix of genes tested versus enrichment tests

enrichIM(Mem): The pathway/P-value matrix

memIM(Mem): The gene-pathway incidence matrix

genes(Mem): List genes represented

thresholds(Mem): List thresholds defined

sets(Mem): List gene sets represented

See also

Examples

# See vignettes for a full walk-through

# list_to_Mem examples
data(Memtest)
mem <- as(Memtest, "list")
Mem <- list_to_Mem(mem)

# examples for Mem_to_list and as(Mem, "list")
data(Memtest)
mem1 <- Mem_to_list(Memtest)
mem2 <- as(Memtest, "list")
identical(mem1, mem2)
#> [1] TRUE