Reese 2019 cohort multienrichment data

Data used for multi-enrichment analysis as described in Reese et al 2019 doi:10.1016/j.jaci.2018.11.043, Epigenome-wide meta-analysis of DNA methylation and childhood asthma. It is provided for convenient testing of multienrichjam.

Usage

Memtest

Reese_genes

Format

Memtest is a Mem multi-enrichment object including two enrichment tests 'Newborns' and 'OlderChildren'

Reese_genes is a list with names 'Newborns' and 'OlderChildren', with a character vectors with genes associated with each cohort.

Source

Reese et al 2019 doi:10.1016/j.jaci.2018.11.043, Epigenome-wide meta-analysis of DNA methylation and childhood asthma

Details

Memtest is a Mem object with enrichment results for 'Newborns' and 'OlderChildren' as described in

See also

Examples

# Mem data for Reese 2019 cohorts
data(Memtest)
# enrichment names
names(Memtest)
#> [1] "NewBorns"   "Older Kids"
# genes represented
genes(Memtest)
#>  [1] "Adcy3"   "Casp8"   "Cited2"  "Ddit4"   "Igf1r"   "Il4"     "Inadl"  
#>  [8] "Ms4a2"   "Napa"    "Nr0b2"   "Polr2b"  "Ppp2ca"  "Prkar2b" "Prkch"  
#> [15] "Prkcz"   "Rbl2"    "Rps23"   "Rps6ka2" "Rptor"   "Slc7a1"  "Smarcd3"
#> [22] "Ywhaq"  
# gene sets, pathways represented
sets(Memtest)
#>  [1] "Huntington's Disease Signaling"                                       
#>  [2] "mTOR Signaling"                                                       
#>  [3] "Role Of Nfat In Cardiac Hypertrophy"                                  
#>  [4] "RAR Activation"                                                       
#>  [5] "Creb Signaling In Neurons"                                            
#>  [6] "Tight Junction Signaling"                                             
#>  [7] "Dopamine-Darpp32 Feedback In Camp Signaling"                          
#>  [8] "Hepatic Cholestasis"                                                  
#>  [9] "eNOS Signaling"                                                       
#> [10] "Production Of Nitric Oxide And Reactive Oxygen Species In Macrophages"
#> [11] "Synaptic Long Term Depression"                                        
#> [12] "Fc Epsilon RI Signaling"                                              
#> [13] "Glioma Signaling"                                                     
#> [14] "HIPPO Signaling"                                                      
#> [15] "Growth Hormone Signaling"                                             
#> [16] "p70S6K Signaling"                                                     

# Reese 2019 genes by cohort
data(Reese_genes)
# Number of genes per cohort
lengths(Reese_genes)
#>      Newborns OlderChildren 
#>            41           162