Apply minimum node spacing for each Cnet node set

Apply minimum node spacing for each Cnet node set

apply_nodeset_spacing(cnet, percent_spacing = 3, apply_negative = FALSE, ...)

Arguments

cnet
igraph object that contains attribute "nodeType" and values nodeType=c("Gene", "Set"), consistent with use of get_cnet_nodesets() and adjust_cnet_nodeset().

Value

igraph Cnet object, whose nodes are adjusted according to the arguments given.

Details

This function automates the process of spacing nodesets in a Cnet plot by calling adjust_cnet_nodeset() on all nodesets whose median nearest node distance is above or below a threshold.

By default nodes must be at least percent_spacing distance apart, relative to the maximum x-axis and y-axis layout coordinate range. Roughly speaking, nodes are clearly visible with distance between nodes of at least 3 percent of the coordinate range.

Nodes are not compressed down to this distance unless using the argument apply_negative=TRUE.

Note that this function does not detect when two clusters overlap. For other specific layout adjustments use adjust_cnet_nodeset(). In future, this function may attempt to re-position nodesets relative to each other, but that process tends to be iterative and is not guaranteed to have a solution.

See also