Find enrichment colnames

Find enrichment colnames

Usage

find_enrich_colnames(
  x,
  keyColname = c("ID", "Name", "pathway", "itemsetID", "Description"),
  nameColname = c("Name", "pathway", "Description", "itemsetID", "ID"),
  descriptionColname = c("Description", "Name", "Pathway", "ID"),
  geneColname = c("geneID", "geneNames", "Genes"),
  countColname = c("gene_count", "count", "geneHits"),
  geneRatioColname = c("GeneRatio", "^Ratio"),
  pvalueColname = c("padjust", "p.adjust", "adjp", "padj", "qvalue", "qval", "q.value",
    "pvalue", "p.value", "pval", "FDR"),
  directionColname = c("activation.z.{0,1}score", "z.{0,1}score"),
  pathGenes = c("setSize", "pathGenes", "Count"),
  geneHits = c("Count", "geneHits", "gene_count"),
  verbose = FALSE,
  ...
)

Arguments

x

data.frame, enrichList, Mem, or list of data.frame objects.

keyColname

character default 'ID' indicating the primary identifier for each set. This column may be a numeric identifier.

nameColname

character default 'Name', with the set name, typically a short name for each set.

descriptionColname

character default 'Description' with the longer set description. It will use nameColname or keyColname as needed.

geneColname

character default 'geneID' containing delimited genes associated with each enrichment result.

countColname

character default 'Count' with the number of genes in the geneColname column. It will be calculated as needed.

geneRatioColname

character default 'GeneRatio' with the numeric (decimal) ratio of test genes to pathway genes, or a character indication in the form '6/24' with 'tested/pathway' gene counts.

pvalueColname

character default 'padjust' with the best available column to use for statistical significance of enrichment.

directionColname

character default 'zscore' with a directional score, typically a z-score or some other reasonably scaled numeric value where the sign indicates directionality, with '+' meaning activated and '-' meaning suppressed.

pathGenes

character default 'setSize' indicating the number of genes in each set as tested for enrichment. This number is not always reported, however it is not used by 'multienrichjam', but is required by some clusterProfiler functions.

geneHits

character default 'Count' indicating the number of genes in the geneColname column. It will be calculated as needed.

verbose

logical default FALSE, whether to print verbose output.

...

additional arguments are ignored.

Value

character of recognized colnames named by the type of column, using NA for any column types not found.

list named by each column name argument, with one character value or NULL in each entry.

See also

Examples

newborn_txt <- system.file("extdata",
   "Newborns-IPA.txt",
   package="multienrichjam");
ipa_dfs <- importIPAenrichment(newborn_txt);
find_enrich_colnames(ipa_dfs[[1]])
#> $keyColname
#> [1] "Name"
#> 
#> $nameColname
#> [1] "Name"
#> 
#> $descriptionColname
#> [1] "Name"
#> 
#> $geneColname
#> [1] "geneNames"
#> 
#> $countColname
#> NULL
#> 
#> $geneRatioColname
#> [1] "Ratio"
#> 
#> $pvalueColname
#> [1] "P-value"
#> 
#> $directionColname
#> NULL
#> 
#> $pathGenes
#> NULL
#> 
#> $geneHits
#> NULL
#> 

er <- enrichDF2enrichResult(ipa_dfs[[1]])
find_enrich_colnames(er)
#> $keyColname
#> [1] "ID"
#> 
#> $nameColname
#> [1] "Description"
#> 
#> $descriptionColname
#> [1] "Description"
#> 
#> $geneColname
#> [1] "geneID"
#> 
#> $countColname
#> [1] "Count"
#> 
#> $geneRatioColname
#> [1] "GeneRatio"
#> 
#> $pvalueColname
#> [1] "p.adjust"
#> 
#> $directionColname
#> NULL
#> 
#> $pathGenes
#> [1] "setSize"
#> 
#> $geneHits
#> [1] "Count"
#>