Return Heatmap column order

Return Heatmap column order as a list of character vectors

heatmap_column_order(hm)

Details

This function is a helpful utility to return the fully qualified list of colnames in a ComplexHeatmap::Heatmap object.

See also