convert signed incidence matrix to list

convert signed incidence matrix to list

imSigned2list(x, verbose = FALSE, ...)

Value

list of named numeric vectors, where list names are defined by colnames(x), and vector names are derived from rownames(x). Values in each vector indicate the signed direction, c(-1,1).

Details

This function converts an signed incidence matrix that contains positive and negative values, or equivalent data.frame, to a list of named vectors containing values c(-1, 1) to indicate signed direction. The input matrix should contain numeric values where positive and negative values indicate directionality. When the input contains only logical values c(TRUE,FALSE) the direction is assumed to be +1 positive.

Values of NA are converted to zero 0 and therefore ignored.

This function uses the transactions class from the arules R package, which in our testing is substantially faster than similar techniques from a variety of other R packages.

See also

Examples

im <- matrix(c(0,1,-1,1,1,NA,-1,0,1),
   ncol=3,
   nrow=3,
   dimnames=list(letters[1:3], LETTERS[1:3]))
print(im);

#>    A  B  C
#> a  0  1 -1
#> b  1  1  0
#> c -1 NA  1

# matrix input
im2list(im);

#> Warning: matrix contains values other than 0 and 1! Setting all entries != 0 to 1.

#> $A
#> [1] "b" "c"
#> 
#> $B
#> [1] "a" "b"
#> 
#> $C
#> [1] "a" "c"
#> 

imSigned2list(im);

#> $A
#>  b  c 
#>  1 -1 
#> 
#> $B
#> a b 
#> 1 1 
#> 
#> $C
#>  c  a 
#>  1 -1 
#> 

imSigned2list(im != 0);

#> $A
#> b c 
#> 1 1 
#> 
#> $B
#> a b 
#> 1 1 
#> 
#> $C
#> a c 
#> 1 1 
#>