Launch ShinyCAT: Cnet Adjustment Tool

Launch ShinyCAT: Cnet Adjustment Tool

Usage

launch_shinycat(g = NULL, envir = new.env(), ..., options = list(width = 1200))

Arguments

g

igraph object, or can be NULL if the variable 'g' is defined in 'envir'.

envir

environment assigned to the R-shiny function space, which may be useful when capturing the igraph object after manipulation via the R-shiny app.

...

additional arguments are ignored.

options

list with additional settings, for example:

host: character string with hostname or IP address to permit incoming connections. Use '0.0.0.0' to accept all incoming host names.

port: numerical port to listen for incoming connections.

quiet: logical whether to

launch.browser: logical whether to launch a web browser, default TRUE for interactive sessions.

display.mode: character string with display mode: 'normal' is recommended; 'showcase' displays the code beside the app.

Value

environment invisibly, containing

'g' the input igraph

'adj_cnet' the output igraph after manipulation in the R-shiny app. It also contains two attributes for reference:

'nodeset_adj': data.frame used to adjust nodesets.

'node_adj': data.frame used to adjust nodes.

Details

This function launches the R-shiny app 'ShinyCat' to manipulate a Cnet igraph object, which is expected to contain node (vertex) attribute 'nodeType' with values 'Gene' and 'Set'.

There are two ways to retain the resulting igraph object:

Capture the output of this function, an environment described below.

Click "Adjustments" then "Save RData", which stores the 'igraph` object as 'adj_cnet'.

The function returns an environment inside which are two objects:

'g': the igraph data input

'adj_cnet': the adjusted Cnet igraph output

Additionally, the 'adj_cnet' object contains two attributes:

'nodeset_adj', 'node_adj': data.frame objects used by bulk_cnet_adjustments() to convert 'g' into 'adj_cnet'.

See also

Examples

# create Cnet test data
g <- make_cnet_test();
cnetenv <- new.env();

# you must catch the output to use the resulting igraph object
# output_envir <- launch_shinycat();