igraph layout function using qgraph Fruchterman-Reingold

igraph layout function using qgraph Fruchterman-Reingold

Usage

layout_with_qfrf(repulse = 3.5, seed = 123, ...)

Arguments

repulse

numeric value typically between 3 and 5, passed to layout_with_qfr(), which in turn is passed to qgraph::qgraph.layout.fruchtermanreingold().

seed

numeric value used to set the R random seed, in order to make layouts consistent.

...

additional arguments are passed to layout_with_qfr().

Value

function used to calculate layout coordinates of an igraph object.

Details

This function returns a layout function, which can be convenient when calling igraph::plot.igraph(), in order to set layout parameters in the same call.

See also