Subgraph using Jam extended logic

Subgraph using Jam extended logic

Usage

subgraph_jam(graph, v)

Arguments

graph

igraph object

v

integer or logical vector indicating the nodes to retain in the final igraph object.

Details

This function extends the igraph::subgraph() function to include proper subset of the graph attribute "layout", which for some unknown reason does not subset the layout matrix consistent with the subset of igraph nodes.

See also