Subset mem multienrichment object

Subset mem multienrichment object

Usage

subset_mem(
  mem,
  includeSets = NULL,
  includeGenes = NULL,
  min_gene_ct = 1,
  min_set_ct = 1,
  p_cutoff = NULL,
  verbose = FALSE,
  ...
)

Arguments

mem

list object as returned by multiEnrichMap()

includeSets

character vector with specific sets to retain, all other sets will be dropped.

includeGenes

character vector with specific genes to retain, all other genes will be dropped.

min_gene_ct

numeric filter applied to genes representing the minimum number of occurrences across sets in the mem$memIM incidence matrix. The default value min_gene_ct=1 effectively requires a gene to be present in at least one set, which is useful after filtering by includeSets.

min_set_ct

numeric filter applied to sets representing the minimum number of occurrences of genes in the mem$memIM incidence matrix. The default value min_set_ct=1 effectively requires a set to contain at least one gene.

p_cutoff

numeric optional enrichment P-value filter to apply to mem$enrichIM enrichment P-values. It is intended to apply optionally higher stringency by using a lower p_cutoff than used by multiEnrichMap().

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

list object of mem data

Details

This function is intended to subset the incidence matrix data contained in a mem object by heuristics. It does not update other data in the mem object such as enrichList and multiEnrichDF, nor any igraph objects. It is intended mainly to subset by sets (pathways), or genes, then also subset other corresponding incidence matrix data consistently.

See also