Subset enrichList for top enrichment results by source

Subset enrichList for top enrichment results by source

topEnrichBySource(enrichDF, n = 15, sourceColnames = c("Category",
  "Source"), sortColname = c("P-value", "-geneHits"),
  newColname = "EnrichGroup", curateFrom = c("CP:.*"),
  curateTo = c("CP"), sourceSubset = NULL, sourceSep = "_",
  descriptionColname = c("Description", "Name", "Pathway"),
  nameColname = c("ID", "Name"), descriptionGrep = NULL,
  nameGrep = NULL, verbose = FALSE, ...)

Arguments

enrichDF
data.frame containing enrichment results.

sourceColnames
character vector of colnames in enrichDF to consider as the "Source". Multiple columns will be combined using delimiter argument sourceSep. When sourceColnames is NULL or contains no colnames(enrichDF), then data is considered "All".

sortColname
character vector indicating the colnames to use to sort data, prior to selecting the top n results by source. This argument is passed to jamba::mixedSortDF(x, byCols=sortColname). Columns can be sorted in reverse order by using the prefix "-", as described in jamba::mixedSortDF().

newColname
new column name to use when sourceColname contains multiple values.

curateFrom, curateTo
character vectors with pattern,replacement values, passed to gsubs() to allow some editing of values. The default values convert MSigDB canonical pathways from the prefix "CP:" to use "CP" which has the effect of combining all canonical pathways before selecting the top n results.

sourceSubset
character vector with a subset of sources to retain. This filter is applied after sourceColname values are combined with delimiter sourceSep if there are multiple colnames in sourceColname.

sourceSep
character string used as a delimiter when sourceColname contains multiple colnames.

descriptionColname, nameColname
character vectors indicating the colnames to consider description and name, as returned from find_colname(). These arguments are used only when descriptionGrep or nameGrep are supplied.

descriptionGrep, nameGrep
character vector of patterns, used to filter pathways to those matching one or more patterns. This argument is used to help extract a specific subset of pathways of interest using keywords. The descriptionGrep argument searches only colname "Description". The nameGrep argument searches only colname "Name".

verbose
logical indicating whether to print verbose output.

Value

data.frame subsetted by the given parameters.

Details

This function takes one enrichResult object, or a data.frame of enrichment results, and determines the top n number of pathways sorted by P-values, within each pathway source.

It is particularly effective with results from MSigDB (http://software.broadinstitute.org/gsea/msigdb/index.jsp) which contains gene sets from a large number of very different sources -- which therefore impart different characteristics on the enrichment results. Notably, certain sources contain small gene sets, which produce enrichment P-values in a different possible range than other sources which provide much larger gene sets.

See also