Subset enrichList for top enrichment results by source

Subset enrichList for top enrichment results by source

topEnrichListBySource(
  enrichList,
  n = 15,
  min_count = 1,
  p_cutoff = 1,
  sourceColnames = c("Category", "Source"),
  sortColname = c("P-value", "pvalue", "padjust", "-GeneRatio", "-geneHits"),
  countColname = c("gene_count", "count"),
  pvalueColname = c("P.Value", "pvalue", "FDR", "adj.P.Val"),
  newColname = "EnrichGroup",
  curateFrom = c("CP:.*"),
  curateTo = c("CP"),
  sourceSubset = NULL,
  sourceSep = "_",
  subsetSets = NULL,
  descriptionColname = c("Description", "Name", "Pathway"),
  nameColname = c("ID", "Name"),
  descriptionGrep = NULL,
  nameGrep = NULL,
  verbose = FALSE,
  ...
)

Arguments

enrichList
list of enrichDF entries, each passed to topEnrichBySource().

n
integer maximum number of pathways to retain, after applying min_count and p_cutoff thresholds if relevant.

min_count
integer minimum number of genes involved in an enrichment result to be retained, based upon values in countColname.

p_cutoff
numeric value indicating the enrichment P-value threshold, pathways with enrichment P-value at or below this threshold are retained, based upon values in pvalueColname.

sourceColnames
character vector of colnames in enrichDF to consider as the "Source". Multiple columns will be combined using delimiter argument sourceSep. When sourceColnames is NULL or contains no colnames(enrichDF), then data is considered "All".

sortColname
character vector indicating the colnames to use to sort data, prior to selecting the top n results by source. This argument is passed to jamba::mixedSortDF(x, byCols=sortColname). Columns can be sorted in reverse order by using the prefix "-", as described in jamba::mixedSortDF().

countColname
character vector of possible colnames in enrichDF that should contain the integer number of genes involved in enrichment. This vector is passed to find_colname() to find an appropriate matching colname in enrichDF.

pvalueColname
character vector of possible colnames in enrichDF that should contain the enrichment P-value used for filtering by p_cutoff.

newColname
new column name to use when sourceColname matches multiple colnames in enrichDF. Values for each row are combined using jamba::pasteByRow().

curateFrom
character vectors with pattern,replacement values, passed to gsubs() to allow some editing of values. The default values convert MSigDB canonical pathways from the prefix "CP:" to use "CP" which has the effect of combining all canonical pathways before selecting the top n results.

curateTo
character vectors with pattern,replacement values, passed to gsubs() to allow some editing of values. The default values convert MSigDB canonical pathways from the prefix "CP:" to use "CP" which has the effect of combining all canonical pathways before selecting the top n results.

sourceSubset
character vector with a subset of sources to retain. If there are multiple colnames in sourceColnames, then column values are combined using jamba::pasteByRow() and delimiter sourceSep, prior to filtering.

sourceSep
character string used as a delimiter when sourceColnames contains multiple colnames.

descriptionColname
character vectors indicating the colnames to consider description and name, as returned from find_colname(). These arguments are used only when descriptionGrep or nameGrep are supplied.

nameColname
character vectors indicating the colnames to consider description and name, as returned from find_colname(). These arguments are used only when descriptionGrep or nameGrep are supplied.

descriptionGrep
character vector of patterns, used to filter pathways to those matching one or more patterns. This argument is used to help extract a specific subset of pathways of interest using keywords. The descriptionGrep argument searches only descriptionColname; the nameGrep argument searches only nameColname.

nameGrep
character vector of patterns, used to filter pathways to those matching one or more patterns. This argument is used to help extract a specific subset of pathways of interest using keywords. The descriptionGrep argument searches only descriptionColname; the nameGrep argument searches only nameColname.

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Details

This function extends topEnrichBySource() by applying consistent filters to each enrichList entry. The purpose of this function is to generate an overall set of pathways to return from any one or more enrichment results. Then apply that filter consistently across all enrichment results.

For example, if "Pathway A" is the top n enriched pathway in enrichList[[1]], then "Pathway A" will be included for every entry in enrichList, even when "Pathway A" is not significant in enrichList[[2]]. The purpose is to allow comparison of enrichment results when a pathway meets the filter criteria in any one entry in enrichList.

See also