Un-nest a nested list into a simple list

Un-nest a nested list into a simple list

unnestList(x, unnamedBase = "x", parentName = NULL, sep = ".",
  makeNameFunc = jamba::makeNames, stopClasses = c("dendrogram",
  "data.frame", "matrix"), ...)

Arguments

x
list potentially containing lists

unnamedBase
character value used as a base for naming any un-named lists, using the format `makeNamesFunc(rep(unnamedBase, n))`.

sep
character delimiter used between nested list names.

stopClasses
vector of classes that should not be un-nested, useful in case some classes inherit list properties.

...
additional arguments are ignored.

makeNamesFunc
function that takes a character vector and returns non-duplicated character vector of equal length. By default it uses `jamba::makeNames()`.

Details

This function inspects a list, and unlists each entry resulting in a simple list of non-list entries as a result.

See also

Examples

L <- list(A=letters[1:10], B=list(C=LETTERS[3:9], D=letters[4:11]));
L;

#> $A
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j"
#> 
#> $B
#> $B$C
#> [1] "C" "D" "E" "F" "G" "H" "I"
#> 
#> $B$D
#> [1] "d" "e" "f" "g" "h" "i" "j" "k"
#> 
#> 

unnestList(L);

#> $A
#>  [1] "a" "b" "c" "d" "e" "f" "g" "h" "i" "j"
#> 
#> $B.C
#> [1] "C" "D" "E" "F" "G" "H" "I"
#> 
#> $B.D
#> [1] "d" "e" "f" "g" "h" "i" "j" "k"
#> 

# inspect the data using str()
str(L);

#> List of 2
#>  $ A: chr [1:10] "a" "b" "c" "d" ...
#>  $ B:List of 2
#>   ..$ C: chr [1:7] "C" "D" "E" "F" ...
#>   ..$ D: chr [1:8] "d" "e" "f" "g" ...

str(unnestList(L))

#> List of 3
#>  $ A  : chr [1:10] "a" "b" "c" "d" ...
#>  $ B.C: chr [1:7] "C" "D" "E" "F" ...
#>  $ B.D: chr [1:8] "d" "e" "f" "g" ...