Mem class

Mem class containing results from multiEnrichMap().

Usage

# S4 method for class 'Mem,ANY,ANY,ANY'
x[i, j, k, drop = FALSE]

list_to_Mem(mem)

Mem_to_list(x, ...)

# S4 method for class 'Mem'
enrichments(x)

# S4 method for class 'Mem'
names(x)

# S4 method for class 'Mem'
enrichments(x) <- value

# S4 method for class 'Mem'
sets(x, ...)

# S4 method for class 'Mem'
sets(x) <- value

# S4 method for class 'Mem'
genes(x, ...)

# S4 method for class 'Mem'
genes(x) <- value

# S4 method for class 'Mem'
geneIM(x, ...)

# S4 method for class 'Mem'
geneIMdirection(x, ...)

# S4 method for class 'Mem'
geneIMcolors(x, ...)

# S4 method for class 'Mem'
enrichList(x, ...)

# S4 method for class 'Mem'
enrichIM(x, ...)

# S4 method for class 'Mem'
enrichIMcolors(x, ...)

# S4 method for class 'Mem'
enrichIMdirection(x, ...)

# S4 method for class 'Mem'
enrichIMgeneCount(x, ...)

# S4 method for class 'Mem'
memIM(x, ...)

# S4 method for class 'Mem'
geneHitIM(x, ...)

# S4 method for class 'Mem'
headers(x)

# S4 method for class 'Mem'
colorV(x, ...)

# S4 method for class 'Mem'
colorV(x) <- value

# S4 method for class 'Mem'
thresholds(x, ...)

# S4 method for class 'Mem'
dim(x)

# S4 method for class 'Mem'
dimnames(x)

Arguments

x

Mem object

i

character or integer with gene names

j

character or integer with pathway names

drop

logical always set to FALSE for Mem objects.

mem

list output from multiEnrichMap()

...

additional arguments are ignored

value

character vector of names to assign to pathway gene sets

Value

list_to_Mem() returns a Mem S4 object, from 'list' or 'Mem' input.

Mem_to_list() returns a list suitable for ther mem functions.

Details

Slots

enrichList: list of enrichResults named by enrichment test.

enrichLabels: character vector with custom enrichment labels, named by enrichment test.

colorV: character vector of R colors, named by enrichment test.

geneHitList: list of vectors named by enrichment test. Vectors are either:
character vector of genes tested as hits for enrichment, or
integer vector named by genes with values indicating the directionality, using 1 or -1.

geneHitIM: numeric matrix with enrichment tests as colnames, and genes as rownames.
Values are 1 or 0 to indicate gene hits tested,
optionally 1 or -1 to indicate directionality.

memIM: numeric matrix with gene sets / pathways as colnames, and genes as rownames.
Values are 1 or 0 indicating genes present in each pathway.

geneIM: numeric matrix with enrichment test as colnames, and genes as rownames.
This matrix differs from 'geneHitIM' in that it only includes genes also in 'memIM', and in that order.

enrichIM: numeric matrix with enrichment test as colnames, and gene sets / pathways as rownames.
Matrix values are enrichment P-values, using the appropriate adjusted P-value or FDR to represent significance.

multiEnrichDF: data.frame with single combined enrichment table.
Data include union of genes involved in each pathway, and lowest P-value observed across all enrichment tests.

multiEnrichResult: enrichResult equivalent of 'multiEnrichDF'.
It is intended to facilitate use in enrichplot functions.

thresholds: list of thresholds used when calling multiEnrichMap().

headers: list to associate actual colnames (headers) in 'enrichList' with the conceptual names used in multienrichjam.

geneIMcolors: character matrix with enrichment test colnames, and gene rownames.
Values are colors assigned using 'colorV'.

geneIMdirection: numeric matrix equivalent to 'geneIM' with values 1 and -1 indicating directionality.
When no directionality is provided, all values will be 1 or 0.

enrichIMcolors: character matrix with enrichment test colnames, and gene set / pathway rownames.
Values are color gradients using the appropriate P-value threshold used with multiEnrichMap().

enrichIMdirection: numeric matrix equivalent to 'enrichIM' with values indicating directional score.
Directional scores may include a IPA z-score or other directional score where available.
When not available, all values are 1.

enrichIMgeneCount: integer matrix equivalent to 'enrichIM'.
Values are the number of genes involved in the enrichment of each pathway.

Functions

x[i: Subset a Mem object, Mem[i, j, k], where i are genes, j are pathways, and k are enrichments.

list_to_Mem(): Convert legacy list mem format to S4 Mem

Mem_to_list(): Coerce S4 Mem to legacy list mem format

enrichments(Mem): Names for the enrichment tests in a Mem object

names(Mem): Names for the enrichment tests in a Mem object

enrichments(Mem) <- value: Assign names for the enrichment tests in a Mem object

sets(Mem): List pathway gene set names in a Mem object

sets(Mem) <- value: Assign pathway gene set names to a Mem object

genes(Mem): List genes represented

genes(Mem) <- value: Assign gene names in a Mem object

geneIM(Mem): The gene-enrichment matrix of genes represented in enrichment tests.

geneIMdirection(Mem): The matrix of genes tested versus enrichment tests, with directionality.

geneIMcolors(Mem): The matrix of colors indicating genes in each enrichment test

enrichList(Mem): The list of enrichResult data in an Mem object

enrichIM(Mem): The pathway/P-value matrix

enrichIMcolors(Mem): The pathway/directional enrichment color matrix

enrichIMdirection(Mem): The pathway-enrichment directional score matrix

enrichIMgeneCount(Mem): Pathway-enrichment gene count matrix, genes involved in enrichment of each pathway, for each enrichment test.

memIM(Mem): The gene-pathway incidence matrix

geneHitIM(Mem): The matrix of genes tested for enrichment, including genes not associated with enrichment results.

headers(Mem): Enrichment data column headers associated to enrichResult data in a Mem object

colorV(Mem): List colors assigned to represented

colorV(Mem) <- value: Assign colors to enrichments in a Mem object. If supplied 'value' matches the number of 'enrichments(x)', colors are assigned in the order supplied. Otherwise, 'names(value)' must be present in 'enrichments(x)', and values are assigned in matching order.

thresholds(Mem): List thresholds defined

dim(Mem): dimensions in order of genes, sets, and enrichments

dimnames(Mem): dimension names for each genes, sets, and enrichments

See also

Examples

# See vignettes for a full walk-through

# list_to_Mem examples
data(Memtest)
mem <- as(Memtest, "list")
Mem <- list_to_Mem(mem)

# examples for Mem_to_list and as(Mem, "list")
data(Memtest)
mem1 <- Mem_to_list(Memtest)
mem2 <- as(Memtest, "list")
identical(mem1, mem2)
#> [1] TRUE