Color edges by nodegroups

Color edges by nodegroups

Usage

color_nodes_by_nodegroups(
  g,
  nodegroups,
  nodegroup_colors = NULL,
  node_alpha = NULL,
  Crange = c(60, 100),
  Lrange = c(45, 85),
  color_attributes = c("color"),
  verbose = TRUE,
  ...
)

Arguments

g

igraph object that contains vertex node attribute "color" as seen with igraph::vertex_attr(g, "color").

nodegroups

list or communities object that references nodes in g and assigns one or more to nodegroups.

verbose

logical indicating whether to print verbose output.

...

additional arguments are passed to colorjam::blend_colors().

edge_alpha

numeric or NULL, where numeric value sets the edge alpha transparency, where edge_alpha=0 is completely transparent, edge_alpha=0.5 is 50% transparent, and edge_alpha=1 is completely not transparent, and is opaque. When edge_alpha=NULL the alpha values are supplied by colorjam::blend_colors() which blends the two values.

Value

igraph object with edge color attribute updated to represent the result of blending node colors, seen by igraph::edge_attr(g)$color.

See also