Convert enrichList to IM incidence matrix

Convert enrichList to IM incidence matrix

enrichList2IM(
  enrichList,
  addAnnotations = TRUE,
  keyColname = c("ID", "Name", "pathway", "itemsetID"),
  valueColname = c("qvalue", "q.value", "pvalue", "p.value"),
  emptyValue = NA,
  verbose = FALSE,
  GmtT = NULL,
  ...
)

Arguments

enrichList

list of enrichResult objects

addAnnotations

logical not implemented, this argument is paired with GmtT.

keyColname

character used to match colnames, referring to the unique identifier. Values in this column will become the column headers in the resulting incidence matrix.

valueColname

character used to match colnames to determine the value to place in each cell of the incidence matrix.

emptyValue

numeric value used to fill empty cells in the incidence matrix. When NULL and valueColname contains "gene", "count", "num", "hit" then emptyValue=0, otherwise emptyValue=1 is used with the assumption that valueColname refers to P-values.

verbose

logical indicating whether to print verbose output.

GmtT

not implemented here, experimental gene set object format that uses arules::transactions class, an efficient object with robust access functions in arules.

...

additional arguments are ignored.

Details

This function takes a list of enrichResult objects and creates an incidence matrix using the value defined by valueColname.

TODO: Port to use venndir::list2im_value(), which itself may be moved to its own proper R package for set and list manipulation, without the dependencies incurred by venndir.

See also