Find colname by character string or pattern matching

Find colname by character string or pattern matching

find_colname(pattern, x, max = 1, index = FALSE, verbose = FALSE,
  ...)

Arguments

pattern
character vector containing text strings or regular expression patterns.

x
input data.frame or other R object that contains colnames.

max
integer maximum number of results to return.

index
logical indicating whether to return the column index as an integer vector. When index=FALSE it returns the matching colnames(x); when index=TRUE it returns the matching column numbers as an integer vector.

...
additional arguments are passed to jamba::provigrep().

Value

character vector with length max, or if no pattern match is found it returns NULL. Also if there are no colnames(x) then it returns NULL.

Details

This function is intended to help find a colname given a character vector of expected values or regular expression patterns. By default it returns the first matching value, but can return multiple if max=Inf.

If there are no colnames(x) then NULL is returned.

The order of operations:

Match exact string.

Match exact string in case-insensitive manner.

Match the start of each string using jamba::provigrep().

Match the end of each string using jamba::provigrep().

Match each string using jamba::provigrep().

The results from the first successful operation is returned.

When there are duplicate colnames(x) only the first unique name is returned.