Get Cnet node set by connected Sets

Get Cnet node set by connected Sets

get_cnet_nodeset(g, set_nodes = NULL, sep = ",", filter_set_only = TRUE, ...)

Arguments

g

igraph object specifically containing Cnet plot data, with node attribute "nodeType" that has values c("Gene", "Set").

set_nodes

character vector of one or more set names, which are defined here as the names of the nodes with nodeType="Set" to which each node with nodeType="Gene" is connected. For example set_nodes=c("A", "B") will return all Gene nodes that are connected only to Set nodes "A" and "B". If set_nodes=NULL then this function returns a list with all set_nodes observed.

sep

character string used as a delimited when combining set_nodes into a name.

...

additional arguments are ignored.

Value

character vector when argument set_nodes is supplied as a character vector; when set_nodes is NULL, this function returns a list of character vectors whose names represent the Set node names connected to each Gene node, where multiple Set node names are delimited by sep.

Details

This function operates on a Cnet igraph object, distinguished by node attribute "nodeType" with value "Gene" for Gene nodes, and "Set" for Set nodes. This function returns Gene nodes that are connected to the Sets given by argument set_nodes. It is useful to identify which genes are connected to Set "A" and "B" for example, in other words the Gene nodes in a specific subcluster of the Cnet igraph.

When set_nodes is NULL, this function returns a list containing character vectors, where each vector represents Gene nodes that are connected to each combination of Set nodes. This option is useful for obtaining all possible Gene subclusters.

This function is also called by adjust_cnet_nodeset() in order to manipulate all nodes in a subcluster as a group, for example calling nudge_igraph_node() on the whole set of nodes.

See also