Jam igraph vectorized plot function

Jam igraph vectorized plot function

jam_plot_igraph(
  x,
  axes = FALSE,
  add = FALSE,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  mark.groups = list(),
  mark.shape = 1/2,
  mark.col = head(colorjam::rainbowJam(max(c(1, length(mark.groups))), alpha = 0.3),
    length(mark.groups)),
  mark.border = head(colorjam::rainbowJam(max(c(1, length(mark.groups))), alpha = 1),
    length(mark.groups)),
  mark.expand = 15,
  pie_to_jampie = TRUE,
  use_shadowText = FALSE,
  verbose = FALSE,
  debug = NULL,
  ...
)

Arguments

x
The graph to plot.

axes
Logical, whether to plot axes, defaults to FALSE.

add
Logical scalar, whether to add the plot to the current device, or delete the device's current contents first.

xlim
The limits for the horizontal axis, it is unlikely that you want to modify this.

ylim
The limits for the vertical axis, it is unlikely that you want to modify this.

mark.groups
A list of vertex id vectors. It is interpreted as a set of vertex groups. Each vertex group is highlighted, by plotting a colored smoothed polygon around and “under” it. See the arguments below to control the look of the polygons.

mark.shape
A numeric scalar or vector. Controls the smoothness of the vertex group marking polygons. This is basically the ‘shape’ parameter of the xspline function, its possible values are between -1 and 1. If it is a vector, then a different value is used for the different vertex groups.

mark.col
A scalar or vector giving the colors of marking the polygons, in any format accepted by xspline; e.g. numeric color ids, symbolic color names, or colors in RGB.

mark.border
A scalar or vector giving the colors of the borders of the vertex group marking polygons. If it is NA, then no border is drawn.

mark.expand
A numeric scalar or vector, the size of the border around the marked vertex groups. It is in the same units as the vertex sizes. If a vector is given, then different values are used for the different vertex groups.

pie_to_jampie
logical indicating whether to convert vertex shape "pie" to "jampie" in order to use vectorized plotting.

...
Additional plotting parameters. See igraph.plotting for the complete list.

Details

This function is a complete copy of igraph:::plot.igraph() with two changes to enable vectorized plotting in these situations:

When there are multiple vertex "shape" attributes, the base plot function draws each individual igraph vertex one by one. The jam_plot_igraph() plots nodes of each shape in a group. For fairly large igraph objects (about 1000 nodes), this change is substantially faster.

When there are multiple font families, the default plot function draws each label one by one. The jam_plot_igraph() draws labels in groups of font family. This situation is very rare, however when used the speed improvement is substantial.

Note that this function is not called by default, and is only called by multienrichjam::jam_igraph().

Smaller changes: mark.col and mark.border default colors now call colorjam::rainbowJam() instead of grDevices::rainbow().

All arguments are documented in igraph::plot.igraph().

See also