convert list to directional incidence matrix

convert list to directional incidence matrix

list2imSigned(x, makeUnique = TRUE, doTimer = FALSE, verbose = FALSE,
  ...)

Value

numeric matrix with rownames defined by names from each vector in the input list, and colnames defined by the names of the input list. Values in the matrix are values from each vector.

Details

This function extends list2im in that it stores the value associated with each element in the list. As such, the input format is a named vector, where the names of the vector are the items, and the numeric values are the values to be stored in the incidence matrix.

A common scenario is to generate a vector of genes, with values c(-1, 0, 1) indicating the direction of gene expression changes, named by the gene symbol. Each vector in the list represents one statistical test. Here, list2imSigned() will convert this list into a directional matrix representing the gene changes across the comparisons.

Examples

L1 <- list(A=c("C","A","B","A"),
   D=c("D","E","F","D"),
   A123=c(1:8,3,5),
   T=LETTERS[7:9]);
L1;

#> $A
#> [1] "C" "A" "B" "A"
#> 
#> $D
#> [1] "D" "E" "F" "D"
#> 
#> $A123
#>  [1] 1 2 3 4 5 6 7 8 3 5
#> 
#> $T
#> [1] "G" "H" "I"
#> 

# Convert each vector to a signed vector
set.seed(123);
L2 <- lapply(L1, function(i){
   i <- unique(i);
   nameVector(sample(c(-1,1), size=length(i), replace=TRUE), i);
});

#> Error in nameVector(sample(c(-1, 1), size = length(i), replace = TRUE),     i): could not find function "nameVector"

L2;

#> Error in eval(expr, envir, enclos): object 'L2' not found

# Convert to signed incidence matrix
list2imSigned(L2);

#> Error in igrepHas("list", class(x)): could not find function "igrepHas"