Convert multiEnrichMap mem output to EnrichmentMap emap

Convert multiEnrichMap mem output to EnrichmentMap emap

mem2emap(
  mem,
  overlap = 0.2,
  p_cutoff = mem$p_cutoff,
  min_count = 4,
  colorV = mem$colorV,
  cluster_function = igraph::cluster_walktrap,
  cluster_list = NULL,
  num_keep_terms = 3,
  keep_terms_sep = ",\n",
  repulse = 3.3,
  remove_singlets = TRUE,
  color_by_nodes = FALSE,
  apply_direction = TRUE,
  direction_max = 2,
  direction_floor = 0.5,
  do_plot = TRUE,
  ...
)

Arguments

mem

list object output from multiEnrichMap()

overlap

numeric value between 0 and 1 indicating the Jaccard overlap coefficient required between any two pathways in order to create a network edge connecting these two pathways. Typically, overlap=0.2 is default, which specifies roughly 20% overlap in genes shared between two pathway nodes. Note these genes must be involved in enrichment, and therefore does not use all possible genes annotated to a pathway. Therefore connections are only created with enriched genes are shared between pathways.

p_cutoff

numeric threshold used for significant enrichment P-value, usually defined in the mem object.

min_count

integer threshold for minimum genes involved in enrichment in order for a pathway to be considered significant during this analysis.

colorV

character vector of R colors used for each enrichment.

cluster_function

function used to cluster nodes in the resulting igraph object, used to help generate a visual summary.

num_keep_terms

integer number of terms to keep from each pathway cluster, when cluster_function is supplied above. Common terms are removed from each pathway cluster, then remaining terms are sorted by decreasing occurrence, and used as a straightforward summary of pathways in each cluster.

keep_terms_sep

character string used to separated multiple pathway terms defined by num_keep_terms above.

repulse

numeric value passed to layout_with_qfr().

remove_singlets

logical indicating whether to remove pathway singlets, which have no connections to any other pathways. It can help simplify busy figures, however removing a singlet pathway is not recommended because it may imply the pathways were not statistically significant, and in fact they were.

color_by_nodes

logical indicating whether to colorize pathway clusters based upon blending the node colors within each cluster. Note that a mix of colors often turns brown, so this feature has unpredictable benefit.

do_plot

logical indicating whether to render the resulting plot.

...

additional arguments are passed to jam_igraph() to customize the network plot.

Details

This function is currently In development.

This function is intended to convert mem output from multiEnrichMap() into an EnrichmentMap igraph format which represents the statistical enrichment support from each pathway enrichment.

This function can apply P-value thresholds using the input mem, or using a custom value.

A node clustering function is applied by default, which may help define suitable subgroups of nodes. When defined, the clusters are used to define nodegroups for edge bundling.

See also