Convert MultiEnrichment incidence matrix to Cnet plot

Convert MultiEnrichment incidence matrix to Cnet plot

memIM2cnet(
  memIM,
  categoryShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  geneShape = c("pie", "coloredrectangle", "circle", "ellipse"),
  categoryColor = "#E5C494",
  geneColor = "#B3B3B3",
  categoryLabelColor = c("blue3"),
  geneLabelColor = "grey40",
  categorySize = 5,
  geneSize = 2.5,
  categoryCex = 0.9,
  geneCex = 0.7,
  frame_darkFactor = 1.4,
  geneIM = NULL,
  geneIMcolors = NULL,
  enrichIM = NULL,
  enrichIMcolors = NULL,
  coloredrect_nrow = 1,
  coloredrect_ncol = NULL,
  coloredrect_byrow = TRUE,
  verbose = FALSE,
  ...
)

Value

igraph object with Concept network data, containing pathways connected to genes. Each node has attribute "nodeType" of either "Set" or "Gene".

Details

This function takes list output from multiEnrichMap() and uses three elements to create a Cnet plot:

"memIM" gene-pathway incidence matrix

"geneIM" the gene incidence matrix

"enrichIM" the pathway enrichment matrix

See also