MultiEnrichment Heatmap of enrichment P-values

MultiEnrichment Heatmap of enrichment P-values

mem_enrichment_heatmap(
  mem,
  p_cutoff = 0.05,
  p_floor = 1e-06,
  row_method = "euclidean",
  name = "-log10P",
  row_dend_reorder = TRUE,
  row_fontsize = 8,
  column_fontsize = 12,
  cluster_columns = FALSE,
  sets = NULL,
  color_by_column = FALSE,
  cex.axis = 1,
  column_title = NULL,
  ...
)

Arguments

mem
list object created by multiEnrichMap(). Specifically the object is expected to contain enrichIM.

p_cutoff
numeric value of the enrichment P-value cutoff, above which P-values are not colored, and are therefore white. The enrichment P-values are displayed as an annotated heatmap at the top of the main heatmap. Any cell that has a color meets at least the minimum P-value threshold.

p_floor
numeric minimum P-value used for the color gradient. P-values below this floor are colored with the maximum color gradient.

row_method
character string of the distance method to use for row and column clustering. The clustering is performed by amap::hcluster().

name
character value passed to ComplexHeatmap::Heatmap(), used as a label above the heatmap color legend.

row_dend_reorder
logical indicating whether to reorder the row dendrogram using the method described in ComplexHeatmap::Heatmap(). The end result is minor reshuffling of leaf nodes on the dendrogram based upon mean signal in each row, which can sometimes look cleaner.

row_fontsize, column_fontsize
arguments passed to ComplexHeatmap::Heatmap() to size row and column labels.

cluster_columns
logical indicating whether to cluster heatmap columns, by default columns are not clustered.

sets
character vector of sets (pathways) to include in the heatmap, all other sets will be excluded.

color_by_column
logical indicating whether to colorize the heatmap using mem$colorV as defined for each comparison. This option is currently experimental, and produces a base R heatmap using jamba::imageByColors().

cex.axis
sent to jamba::imageByColors() only when color_by_column=TRUE.

column_title
optional character string with title to display above the heatmap.

...
additional arguments are passed to ComplexHeatmap::Heatmap() for customization.

Details

This function is a lightweight wrapper to ComplexHeatmap::Heatmap() intended to visualize the enrichment P-values from multiple enrichment results. The P-value threshold is used to colorize every cell whose P-value meets the threshold, while all other cells are therefore white.

See also