MultiEnrichment Heatmap of Genes and Pathways

MultiEnrichment Heatmap of Genes and Pathways

mem_gene_path_heatmap(
  mem,
  genes = NULL,
  sets = NULL,
  min_gene_ct = 2,
  min_set_ct = 3,
  column_fontsize = NULL,
  row_fontsize = NULL,
  row_method = "binary",
  column_method = "binary",
  cluster_columns = NULL,
  cluster_rows = NULL,
  name = "gene_ct",
  p_cutoff = 0.05,
  row_split = NULL,
  column_split = NULL,
  auto_split = TRUE,
  column_title = NULL,
  row_title = NULL,
  row_title_rot = 90,
  verbose = FALSE,
  ...
)

Arguments

mem
list object created by multiEnrichMap(). Specifically the object is expected to contain colorV, enrichIM, memIM, geneIM.

genes
character vector of genes to include in the heatmap, all other genes will be excluded.

sets
character vector of sets (pathways) to include in the heatmap, all other sets will be excluded.

min_gene_ct
minimum number of occurrences of each gene across the pathways, all other genes are excluded.

min_set_ct
minimum number of genes required for each set, all other sets are excluded.

column_fontsize, row_fontsize
arguments passed to ComplexHeatmap::Heatmap() to size column and row labels.

row_method, column_method
character string of the distance method to use for row and column clustering. The clustering is performed by amap::hcluster().

name
character value passed to ComplexHeatmap::Heatmap(), used as a label above the heatmap color legend.

p_cutoff
numeric value of the enrichment P-value cutoff, above which P-values are not colored, and are therefore white. The enrichment P-values are displayed as an annotated heatmap at the top of the main heatmap. Any cell that has a color meets at least the minimum P-value threshold.

column_title
optional character string with title to display above the heatmap.

...
additional arguments are passed to ComplexHeatmap::Heatmap() for customization.

Details

This function takes the mem list output from multiEnrichMap() and creates a gene-by-pathway incidence matrix heatmap, using ComplexHeatmap::Heatmap(). It uses three basic sources of data to annotate the heatmap:

mem$memIM the gene-set incidence matrix

mem$geneIM the gene incidence matrix by dataset

mem$enrichIM the pathway enrichment P-value matrix by dataset

It will try to estimate a reasonable number of column and row splits in the dendrogram, based solely upon the number of columns and rows. These guesses can be controlled with argument column_split and row_split, respectively.

See also