Multienrichment folio of summary plots

Multienrichment folio of summary plots

mem_plot_folio(
  mem,
  p_cutoff = NULL,
  main = "",
  p_floor = 1e-06,
  use_raster = TRUE,
  min_gene_ct = 2,
  min_set_ct = 2,
  column_method = "euclidean",
  row_method = "euclidean",
  exemplar_range = c(1, 2, 3),
  pathway_column_split = NULL,
  pathway_row_split = NULL,
  edge_color = NULL,
  cex.main = 2,
  cex.sub = 1.5,
  do_which = NULL,
  verbose = TRUE,
  ...
)

Arguments

mem
list object created by multiEnrichMap(). Specifically the object is expected to contain colorV, enrichIM, memIM, geneIM.

p_cutoff
numeric value indicating the enrichment P-value threshold used for multiEnrichMap(), but when NULL this value is taken from the mem input, or 0.05 is used by default.

main
character string used as a title on Cnet plots.

p_floor
numeric value indicating the lowest enrichment P-value used in the color gradient on the Enrichment Heatmap.

use_raster
logical indicating whether to use raster heatmaps, passed to ComplexHeatmap::Heatmap().

min_gene_ct, min_set_ct
integer values passed to mem_gene_pathway_heatmap(). The min_gene_ct requires each set to contain min_gene_ct genes, and min_set_ct requires each gene to be present in at least min_set_ct sets.

column_method, row_method
arguments passed to ComplexHeatmap::Heatmap() which indicate the distance method used to cluster columns and rows, respectively.

exemplar_range
integer vector (or NULL) used to create Cnet exemplar plots, using this many exemplars per cluster.

pathway_column_split
integer value passed as column_split to mem_gene_path_heatmap(), indicating the number of pathway clusters to create in the gene-pathway heatmap.

cex.main, cex.sub
numeric values passed to title() which size the default title and sub-title in Cnet plots.

do_which
integer vector of plots to produce. When do_which is NULL, then all plots are produced. This argument is intended to help produce one plot from a folio, therefore each plot is referred by the number of the plot, in order.

...
additional arguments are passed to downstream functions.

Value

list is returned via invisible, which contains each relevant object.

Details

This function is intended to create multiple summary plots from a Multienrichment result, the list output from multiEnrichMap().

The plots created:

Heatmap using enrichment P-values using mem_enrichment_heatmap()

Gene-Pathway Heatmap using mem_gene_pathway_heatmap()

Cnet plot using Gene-Pathway Heatmap collapsed clusters

Cnet plot using Gene-Pathway Heatmap cluster examplars (n per cluster)

Cnet plots one per cluster

See also