Multienrichment folio of summary plots

Multienrichment folio of summary plots

mem_plot_folio(
  mem,
  do_which = NULL,
  p_cutoff = NULL,
  p_floor = 1e-06,
  main = "",
  use_raster = TRUE,
  min_gene_ct = 2,
  min_set_ct = 2,
  min_set_ct_each = 4,
  column_method = "euclidean",
  row_method = "euclidean",
  exemplar_range = c(1, 2, 3),
  pathway_column_split = NULL,
  pathway_column_title = NULL,
  gene_row_split = NULL,
  gene_row_title = NULL,
  edge_color = NULL,
  cex.main = 2,
  cex.sub = 1.5,
  row_cex = 1,
  column_cex = 1,
  max_labels = 4,
  max_nchar_labels = 25,
  include_cluster_title = TRUE,
  repulse = 4,
  use_shadowText = FALSE,
  color_by_column = FALSE,
  enrich_im_weight = 0.3,
  gene_im_weight = 0.5,
  colorize_by_gene = TRUE,
  cluster_color_min_fraction = 0.4,
  byCols = c("composite_rank", "minp_rank", "gene_count_rank"),
  do_plot = TRUE,
  verbose = TRUE,
  ...
)

Arguments

mem
list object created by multiEnrichMap(). Specifically the object is expected to contain colorV, enrichIM, memIM, geneIM.

do_which
integer vector of plots to produce. When do_which is NULL, then all plots are produced. This argument is intended to help produce one plot from a folio, therefore each plot is referred by the number of the plot, in order.

p_cutoff
numeric value indicating the enrichment P-value threshold used for multiEnrichMap(), but when NULL this value is taken from the mem input, or 0.05 is used by default.

p_floor
numeric value indicating the lowest enrichment P-value used in the color gradient on the Enrichment Heatmap.

main
character string used as a title on Cnet plots.

use_raster
logical indicating whether to use raster heatmaps, passed to ComplexHeatmap::Heatmap().

min_gene_ct, min_set_ct
integer values passed to mem_gene_path_heatmap(). The min_gene_ct requires each set to contain min_gene_ct genes, and min_set_ct requires each gene to be present in at least min_set_ct sets.

min_set_ct_each
minimum number of genes required for each set, required for at least one enrichment test.

column_method, row_method
arguments passed to ComplexHeatmap::Heatmap() which indicate the distance method used to cluster columns and rows, respectively.

exemplar_range
integer vector (or NULL) used to create Cnet exemplar plots, using this many exemplars per cluster.

pathway_column_split, gene_row_split
integer value passed as column_split and row_split, respectively, to mem_gene_path_heatmap(), indicating the number of pathway clusters, and gene clusters, to create in the gene-pathway heatmap. When either value is NULL then auto-split logic is used.

pathway_column_title, gene_row_title
character vectors passed to mem_gene_path_heatmap() as column_title and row_title, respectively. When one value is supplied, it is displayed and centered across all the respective splits. When multiple values are supplied, values are used to the number of splits, and recycled as needed. In that case, repeated values are made unique by jamba::makeNames().

cex.main, cex.sub
numeric values passed to title() which size the default title and sub-title in Cnet plots.

row_cex, column_cex
numeric character expansion factor, used to adjust the relative size of row and column labels, respectively. A value of 1.1 will make row font size 10% larger.

color_by_column
logical indicating whether to colorize the enrichment heatmap columns using colorV in the input mem. This argument is only relevant when do_which include 1.

enrich_im_weight, gene_im_weight
numeric value between 0 and 1, passed to mem_gene_path_heatmap(), used to apply relative weight to clustering columns and rows, respectively, when combining the gene-pathway incidence matrix with either column enrichment P-values, or row gene incidence matrix data.

colorize_by_gene
logical passed to mem_gene_path_heatmap() indicating whether the heatmap body for the gene-pathway heatmap will be colorized using the enrichment colors for each gene.

cluster_color_min_fraction
numeric value passed to collapse_mem_clusters() used to determine which enrichment colors to associate with each Cnet cluster.

byCols
character vector describing how to sort the pathways within Cnet clusters. This argument is passed to rank_mem_clusters().

do_plot
logical indicating whether to render each plot. When do_plot=FALSE the plot objects will be created and returned, but the plot itself will not be rendered. This option may be useful to generate the full set of figures in one set, then review each figure one by one in an interactive session.

verbose
logical indicating whether to print verbose output.

...
additional arguments are passed to downstream functions. Notably, sets is passed to mem_gene_path_heatmap() which allows one to define a specific subset of sets to use in the gene-pathway heatmap.

Value

list is returned via invisible, which contains each plot object enabled by the argument do_which:

enrichment_hm is a Heatmap object from ComplexHeatmap that contains the enrichment P-value heatmap. Note that this data is not used directly in subsequent plots, the pathway clusters shown here are based upon -log10(Pvalue) and not the underlying gene content of each pathway. This plot is a useful overview that answers the question "How many pathways are significantly enriched across the different enrichment tests?"

gp_hm is a Heatmap object from ComplexHeatmap with the gene-pathway incidence matrix heatmap. This heatmap and the column/pathway clusters are the subject of subsequent Cnet plots.

gp_hm_caption is a text caption that describes the gene and set filter criteria, and the row and column distance methods used for clustering. Because the filtering and clustering options have substantial impact on clustering, and the pathway clusters are the key for all subsequent plots, these values are important to keep associated with the output of this function.

clusters_mem is a list with the pathways contained in each pathway cluster shown by the gene-pathway heatmap, obtained by heatmap_column_order(gp_hm). The pathway names should also be present in colnames(mem$memIM) and rownames(mem$enrichIM), for follow-up inspection.

cnet_collapsed is an igraph object with Cnet plot data, where the pathways have been collapsed by cluster, using the gene-pathway heatmap clusters defined in clusters_mem. Each pathway cluster is labeled by cluster name, and the first few pathway names. This data can be plotted using jam_igraph(cnet_collapsed).

cnet_collapsed_set is the same as cnet_collapsed except the pathways are labeled by the cluster name only, for example c("A", "B", "C", "D"). This data can be plotted using jam_igraph(cnet_collapsed_set).

cnet_collapsed_set2 is the same as cnet_collapsed_set except the gene labels are hidden, useful when there are too many genes to label clearly. The gene symbols are still stored in V(g)$name but the labels in V(g)$label are updated to hide the genes. This data can be plotted using jam_igraph(cnet_collapsed_set2).

cnet_exemplars is a list of igraph Cnet objects, each one contains only the number of exemplar pathways from each cluster defined by argument exemplar_range. By default it uses 1 exemplar per cluster, then 2 exemplars per cluster, then 3 exemplars per cluster. A number of published figures use 1 exemplar per pathway cluster. This data can be plotted using jam_igraph(cnet_exemplars[[1]]), which will plot only the first igraph object from the list.

cnet_clusters is a list of igraph Cnet objects, each one contains all the pathways in one pathway cluster. This data can be plotted using jam_igraph(cnet_clusters[[1]]), or by calling a specific cluster jam_igraph(cnet_clusters[["A"]]).

Details

This function is intended to create multiple summary plots using the output data from multiEnrichMap(). By default it creates all plots one by one, sufficient for including in a multi-page PDF document with cairo_pdf(..., onefile=TRUE). However, each plot object can be created and viewed later by using do_plot=FALSE. This function returns a list with each plot object, described in detail below.

The plots created:

Heatmap using enrichment P-values using mem_enrichment_heatmap()

Gene-Pathway Heatmap using mem_gene_path_heatmap()

Cnet plot using Gene-Pathway Heatmap collapsed clusters

Cnet plot using Gene-Pathway Heatmap cluster examplars (n per cluster)

Cnet plots one per cluster

By far the most commonly used plots are do_which=2 for the gene-pathway heatmap, and do_which=4 for the collapsed Cnet plot, where Cnet clusters are based upon the gene-pathway heatmap.

The key step in the workflow is the gene-pathway incidence matrix heatmap. This step also clusters the pathways using a combination of the pathway-gene incidence matrix results, and a weighted matrix of enrichment -log10(Pvalue) from mem$enrichIM. The pathway clusters are used in subsequent Cnet plots. Our experience is that the pathway clustering does not need to be perfect to be useful and valid. The pathway clusters are valid based upon the parameters used for clustering, and provide insight into the genes that help define each cluster distinct from other clusters. Sometimes with clustering techniques, the results are more or less effective based upon the type of pattern observed in the data, so it can be helpful to adjust parameters to drill down to the most effective patterns.

The balance of weighting between the incidence matrix and enrichment P-values is controlled by enrich_im_weight: where enrich_im_weight=0 will only cluster pathways using the enrichment P-values, and enrich_im_weight=1 will only cluster pathways using the gene-pathway incidence matrix. The default enrich_im_weight=0.3 is the recommended starting value based upon our experience.

The argument column_method defines the distance method, for example "euclidean" and "binary" are two immediate choices. The method also adds "correlation" from amap::hcluster() which can be very useful especially with large datasets.

The number of pathway clusters can be controlled with pathway_column_split, by default when pathway_column_split=NULL and auto_cluster=TRUE the number of clusters is defined based upon the total number of pathways. In practice, pathway_column_split=4 is almost always a good choice, partly because this number of clusters is the easiest to visualize in a Cnet plot.

To define your own pathway cluster names, define pathway_column_title as a vector with length equal to pathway_column_split.

The pathway clusters are dependent upon the genes and pathways used during clustering, which are also controlled by min_set_ct and min_gene_ct.

Pathways are also filtered with min_set_ct_each which requires each pathway to contain at least this many genes from at least one enrichment test. Said another way, if min_set_ct_each=4 it requires each pathway to contain 4 or more genes, and at least 4 genes must be present in the same enrichment test as defined in mem$geneIM. If there are 4 genes but each gene was present in only one enrichment test, this pathway would be filtered out.

The resulting Cnet pathway clusters are single nodes in the network, and these nodes are colorized based upon the enrichment tests involved. The threshold for including the color for each enrichment test is defined by cluster_color_min_fraction, which requires at least this fraction of pathways in a pathway cluster meets the significance criteria for that enrichment test. To reduce this filter to include any enrichment test with at least one significant result, use cluster_color_min_fraction=0.01. In the gene-pathway heatmap, these colors are shown across the top of the heatmap. The default cluster_color_min_fraction=0.4 requires 40% of pathways in a cluster for each enrichment test.

See also