Rank Multienrichment clusters

Rank Multienrichment clusters

Usage

rank_mem_clusters(
  mem,
  clusters,
  choose = NULL,
  per_cluster = Inf,
  byCols = c("composite_rank", "minp_rank", "gene_count_rank"),
  verbose = FALSE,
  ...
)

Arguments

mem

Mem or legacy list mem output from multiEnrichMap()

clusters

list containing set names, that must match colnames(mem$memIM) and rownames(mem$enrichIM).

choose

optional vector indicating which clusters to return. If an integer vector, it refers to the elements in the input clusters. If a character vector, it must contain values in names(clusters). When choose is NULL, all clusters are returned.

per_cluster

integer vector with the number of entries to return for each cluster. Values will be recycled to the length of the clusters to be returned, defined by choose or by length(clusters) when choose is NULL.

byCols

character vector used to sort the resulting data.frame within each cluster. This argument is passed directly to jamba::mixedSortDF().

verbose

logical indicating whether to print verbose output.

...

additional arguments are ignored.

Value

data.frame sorted by the criteria defined by byCols, with colname "set" to indicate the pathway/set name.

Details

This function takes list output from multiEnrichMap(), and a list of clusters, and returns a data.frame that contains several rank order metrics. It is intended to be used with column clusters following mem_gene_path_heatmap(), see examples.

The argument per_cluster is intended to make it convenient to pick the top exemplar pathways, especially when argument byCols is defined so that it sorts by the rank columns.

The argument choose is intended to make it easy to retrieve pathways from specific clusters.

See also

Examples

## Start with mem
# mem <- multiEnrichMap(...);
# gp_hm <- mem_gene_path_heatmap(mem, column_split=4);
## Retrieve clusters from the Heatmap output, there should be 4 clusters
# clusters <- heatmap_column_order(gp_hm)
# clusters_df <- rank_mem_clusters(mem, clusters)