Remove igraph singlet nodes

Remove igraph singlet nodes

removeIgraphSinglets(g, min_degree = 1, ...)

Arguments

g
igraph object

min_degree
numeric threshold with the minimum number of connections, also known as the "degree", required for each node.

...
additional arguments are ignored.

Details

This function is a lightweight method to remove igraph nodes with no connections. In fact, the min_degree can be used to require a minimum number of connections, but the intended use is to remove the singlet nodes that have no connections.

See also