Re-order igraph nodes

Re-order igraph nodes

reorderIgraphNodes(
  g,
  sortAttributes = c("pie.color", "coloredrect.color", "color", "label", "name"),
  nodeSortBy = c("x", "y"),
  layout = NULL,
  verbose = FALSE,
  ...
)

Arguments

g
igraph object

sortAttributes
character vector of node attribute names, to be applied in order when sorting nodes.

nodeSortBy
character vector containing "x" and "y" indicating the primary axis used to sort nodes.

layout
numeric matrix of node coordinates, or function used to produce layout coordinates. When layout is NULL, this function tries to use graph attribute "layout", otherwise the relayout_with_qfr is called.

verbose
logical indicating whether to print verbose output.

...
additional arguments are ignored.

Value

igraph with nodes positioned to order nodes by color. The layout coordinates are stored in the graph attribute "layout", accessible with g$layout or graph_attr(g, "layout"). When there are not multiple nodes sharing the same neighbors, the original igraph object is returned, with the addition of layout coordinates.

Details

This function takes an igraph and a layout in the form of coordinates, or a function used to produce coordinates. It repositions nodes within equivalent positions, ordering nodes by color along either the "x" or "y" direction.

Equivalent node positions are those with the same neighboring nodes. For example if node "A" and node "B" both have neighbors c("D", "E", "F") then nodes "A" and "B" are considered equivalent, and will be reordered by their color.

This function is particularly effective with concept network (Cnet) graphs, where multiple terms may be connnected to the same concept. For MultiEnrichmap, it typically works when multiple genes are connected to the same pathways. When this happens, the genes are sorted to group the colors.

See also