plot function for igraph vertex shape coloredrectangle

plot function for igraph vertex shape coloredrectangle

shape.coloredrectangle.plot(coords, v = NULL, params)

Arguments

coords
two-column numeric matrix with x- and y-coordinates, respectively.

v
optional ids of the vertices to plot. It should match the number of rows in the coords argument.

params
a function object that can be called to query vertex/edge/plot graphical parameters. The first argument of the function is 'vertex', 'edge' or 'plot' to decide the type of the parameter, the second is a character string giving the name of the parameter.

Value

invisible list of data.frame objects which were used to draw the rectangle objects. However the purpose of this function is the by-product that it draws rectangles onto an igraph graph.

Details

This function defines the plotting function for custom igraph vertex shape coloredrectangle. The coloredrectangle shape is described as:

a vertex drawn as a rectangle, filled with squares which are each individually colored.

the squares are arrayed into a number of columns and rows, which are defined for each vertex using coloredrect.ncol and coloredrect.nrow, respectively.

the vector of colors is arrayed as values of a matrix, therefore coloredrect.byrow is logical to indicate whether colors should fill the matrix by row, similar to how byrow=TRUE is used.

the colors for each vertex are defined by coloredrect.color. When this value does not exist, this function will attempt to use values from pie.color if they exist.

the size of the rectangle is defined by size2, which is typically the height of a rectangle node. The intent was to use a size parameter which is already convenient, but which does not conflict with the size used for vertex shapes such as "circle".

when a vertex has 3 columns, and 2 rows, and only 5 colors, the colors are not cycled to fill the complete node. Instead, the last color is "transparent" to render no color in that square.

the frame around each node is described with frame.color and frame.lwd which draws one rectangular border around the full vertex. The border around all coloredrect vertices is drawn before the squares are drawn for all coloredrect vertices, so that the border will not overlap the squares. It has the visible effect of showing vertex borders behind the vertex colors, while allowing for vectorized drawing, which is substantially more efficient than per-vertex rendering. The graphics::symbols() function is used to render rectangles, it accepts only one value for lwd and lty, so rectangles are split by lwd and lty and rendered in order. As a result, if each vertex frame had a different lwd,lty combination, the rendering would effectively be non-vectorized. For large numbers of vertices and distinct lwd values, it would be preferred to reduce the lwd values to limited significant digits (e.g. with signif(...,digits=2)). That said, line width is not an optimal way to convey a quantitative measurement.

a colored border can be drawn around each square in the vertex, defined by coloredrect.border and coloredrect.lwd, however it is recommended to use this color only as a visible break, for example the default is "grey30" which draws a simple border. Note that squares are rendered after the vertex frame color, so each square color border will be drawn over (on top of) the frame border.

when coloredrect.ncol does not exist, it will attempt to use two rows of colors if coloredrect.color contains two or more values.

when coloredrect.nrow does not exist, it will use a value based upon coloredrect.ncol and the length of coloredrect.color.

Currently, the plotting of coloredrectangle vertices does not define clipping, which means edges are drawn to the center of each vertex, and the coloredrectangle vertex shapes are drawn on top of the edges. Transparent nodes will therefore show edges beneath them.

See also